feat(app): added Verax dashboard customization
