+++ app/components/NetworkCard.tsx
@@ -1,77 +1,12 @@
-import { Stamp } from "@gitcoin/passport-types";
-import { useContext, useEffect, useState } from "react";
-import { CeramicContext, AllProvidersState, ProviderState } from "../context/ceramicContext";
-import { OnChainContext, OnChainProviderType } from "../context/onChainContext";
-import { ScorerContext, ScoreStateType } from "../context/scorerContext";
+import React, { useContext } from "react";
+import { OnChainContext } from "../context/onChainContext";
 import { SyncToChainButton } from "./SyncToChainButton";
 import { Chain } from "../utils/chains";
-import { FeatureFlags } from "../config/feature_flags";
-
-export enum OnChainStatus {
-  NOT_MOVED,
-  MOVED_OUT_OF_DATE,
-  MOVED_UP_TO_DATE,
-}
-
-type ProviderWithStamp = ProviderState & { stamp: Stamp };
-
-export const checkOnChainStatus = (
-  allProvidersState: AllProvidersState,
-  onChainProviders: OnChainProviderType[],
-  rawScore: number,
-  scoreState: ScoreStateType,
-  onChainScore: number
-): OnChainStatus => {
-  if (onChainProviders.length === 0) return OnChainStatus.NOT_MOVED;
-
-  if (scoreState === "DONE" && rawScore !== onChainScore) return OnChainStatus.MOVED_OUT_OF_DATE;
-
-  const verifiedDbProviders: ProviderWithStamp[] = Object.values(allProvidersState).filter(
-    (provider): provider is ProviderWithStamp => provider.stamp !== undefined
-  );
-
-  const [equivalentProviders, differentProviders] = verifiedDbProviders.reduce(
-    ([eq, diff], provider): [ProviderWithStamp[], ProviderWithStamp[]] => {
-      const expirationDateSeconds = Math.floor(new Date(provider.stamp.credential.expirationDate).valueOf() / 1000);
-      const issuanceDateSeconds = Math.floor(new Date(provider.stamp.credential.issuanceDate).valueOf() / 1000);
-
-      const isEquivalent = onChainProviders.some(
-        (onChainProvider) =>
-          onChainProvider.providerName === provider.stamp.provider &&
-          onChainProvider.credentialHash === provider.stamp.credential.credentialSubject?.hash &&
-          Math.floor(onChainProvider.expirationDate.valueOf() / 1000) === expirationDateSeconds &&
-          Math.floor(onChainProvider.issuanceDate.valueOf() / 1000) === issuanceDateSeconds
-      );
-      return isEquivalent ? [[...eq, provider], diff] : [eq, [...diff, provider]];
-    },
-    [[], []] as [ProviderWithStamp[], ProviderWithStamp[]]
-  );
-
-  return equivalentProviders.length === onChainProviders.length && differentProviders.length === 0
-    ? OnChainStatus.MOVED_UP_TO_DATE
-    : OnChainStatus.MOVED_OUT_OF_DATE;
-};
+import { useOnChainStatus } from "../hooks/useOnChainStatus";
 
 export function NetworkCard({ chain }: { chain: Chain }) {
-  const { allProvidersState } = useContext(CeramicContext);
-  const { onChainProviders, onChainScores, onChainLastUpdates } = useContext(OnChainContext);
-  const { rawScore, scoreState } = useContext(ScorerContext);
-  const [onChainStatus, setOnChainStatus] = useState<OnChainStatus>(OnChainStatus.NOT_MOVED);
-
-  useEffect(() => {
-    const checkStatus = async () => {
-      const savedNetworkProviders = onChainProviders[chain.id] || [];
-      const stampStatus = checkOnChainStatus(
-        allProvidersState,
-        savedNetworkProviders,
-        rawScore,
-        scoreState,
-        onChainScores[chain.id]
-      );
-      setOnChainStatus(stampStatus);
-    };
-    checkStatus();
-  }, [allProvidersState, chain.id, onChainProviders, onChainScores, rawScore, scoreState]);
+  const onChainStatus = useOnChainStatus({ chain });
+  const { onChainLastUpdates } = useContext(OnChainContext);
 
   return (
     <div className="mb-6 rounded border border-foreground-6 bg-background-4 p-0 text-color-2">
